chore(plugins): rename models to plugins
