+++ src/plugins/index.ts
