+++ src/plugins/time-sync.ts
